Integration fixes working

+++ packages/contracts/Interfaces/IDonation.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.20;
 
 interface IDonation {
-    function donate(uint128 amount) external;
+    function donate(uint256 amount) external;
 }
